consolidate-locks: extract lock logic into decorators
