+++ helm/operator/values.yaml
@@ -54,7 +54,7 @@ operator:
   interval: 60
 
   ## @param operator.initial_delay [default: 60] The initial delay in seconds before the first check for changes in the secrets in the pass store.
-  initial_delay: 60
+  initial_delay: 2
 
   ## @param operator.priority [default: 100] The priority of the operator. The higher the number, the higher the priority. Only useful if multiple operators are running.
   priority: 100
